ejerciciosMejorados/p8/requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Django==3.1.2
djangorestframework==3.12.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.19.3
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.2
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
wikipedia==1.4.0
wrapt==1.12.1
